refactor: move charge code into charge/ folder, restructure shared utilities
- Move sdk/src/Methods.ts → sdk/src/charge/Methods.ts
- Move sdk/src/client/ → sdk/src/charge/client/
- Move sdk/src/server/ → sdk/src/charge/server/
- Move sdk/src/scval.ts → sdk/src/shared/scval.ts
- Remove sdk/src/signers.ts (replaced by shared/keypairs.ts)
- Create sdk/src/charge/index.ts
- Update package.json exports: ./client → ./charge/client, ./server → ./charge/server
- Bump version to 0.2.0
- Update all import paths in charge, channel, and example files
