+++ examples/server.ts
@@ -20,8 +20,8 @@ import { join } from 'node:path'
 import { Keypair } from '@stellar/stellar-sdk'
 import { Mppx } from 'mppx/server'
 import { Mppx as MppxClient } from 'mppx/client'
-import { stellar } from '../sdk/src/server/index.js'
-import { stellar as stellarClient } from '../sdk/src/client/index.js'
+import { stellar } from '../sdk/src/charge/server/index.js'
+import { stellar as stellarClient } from '../sdk/src/charge/client/index.js'
 import { USDC_SAC_TESTNET } from '../sdk/src/constants.js'
 import { Env } from './config/charge-server.js'
