+++ sdk/src/channel/server/index.ts
@@ -4,5 +4,5 @@ export type { ChannelState } from './State.js'
 export { stellar } from './Methods.js'
 export { watchChannel } from './Watcher.js'
 export type { ChannelEvent } from './Watcher.js'
-export { resolveKeypair } from '../../signers.js'
+export { resolveKeypair } from '../../shared/keypairs.js'
 export { Expires, Mppx, Store } from 'mppx/server'
