+++ sdk/src/charge/server/index.ts
@@ -1,4 +1,4 @@
 export { charge } from './Charge.js'
 export { stellar } from './Methods.js'
-export { resolveKeypair } from '../signers.js'
+export { resolveKeypair } from '../../shared/keypairs.js'
 export { Expires, Mppx, Store } from 'mppx/server'
