+++ sdk/src/index.ts
@@ -1,5 +1,8 @@
-export * as Methods from './Methods.js'
+// Schemas
+export * as ChargeMethods from './charge/Methods.js'
 export * as ChannelMethods from './channel/Methods.js'
+
+// Constants (public)
 export {
   DEFAULT_DECIMALS,
   DEFAULT_FEE,
@@ -12,8 +15,18 @@ export {
   USDC_SAC_TESTNET,
   XLM_SAC_MAINNET,
   XLM_SAC_TESTNET,
+  ALL_ZEROS,
   type NetworkId,
 } from './constants.js'
-export { fromBaseUnits, toBaseUnits } from './Methods.js'
-export { resolveKeypair } from './signers.js'
+
+// Unit conversion (public, moved from Methods.ts)
+export { fromBaseUnits, toBaseUnits } from './shared/units.js'
+
+// Keypair resolution (public)
+export { resolveKeypair } from './shared/keypairs.js'
+
+// Env parsing (public)
 export * as Env from './env.js'
+
+// Logger interface (public — consumers need the type)
+export type { Logger } from './shared/logger.js'
