+++ sdk/src/shared/scval.ts
